refactor: move C preprocessor-related implementations to own package
