+++ src/main/java/org/variantsync/diffdetective/datasets/predefined/MarlinCPPDiffLineFormulaExtractor.java
@@ -1,6 +1,6 @@
 package org.variantsync.diffdetective.datasets.predefined;
 
-import org.variantsync.diffdetective.feature.CPPDiffLineFormulaExtractor;
+import org.variantsync.diffdetective.feature.cpp.CPPDiffLineFormulaExtractor;
 
 import java.util.regex.Pattern;
 
@@ -8,6 +8,7 @@ import java.util.regex.Pattern;
  * Extracts formulas from preprocessor annotations in the marlin firmware.
  * In particular, it resolves the 'ENABLED' and 'DISABLED' macros that are used in Marlin
  * to check for features being (de-)selected.
+ *
  * @author Paul Bittner
  */
 public class MarlinCPPDiffLineFormulaExtractor extends CPPDiffLineFormulaExtractor {
@@ -17,9 +18,9 @@ public class MarlinCPPDiffLineFormulaExtractor extends CPPDiffLineFormulaExtract
     @Override
     protected String resolveFeatureMacroFunctions(String formula) {
         return
-            replaceAll(ENABLED_PATTERN, "$1",
-                replaceAll(DISABLED_PATTERN, "!($1)",
-                    super.resolveFeatureMacroFunctions(formula)));
+                replaceAll(ENABLED_PATTERN, "$1",
+                        replaceAll(DISABLED_PATTERN, "!($1)",
+                                super.resolveFeatureMacroFunctions(formula)));
     }
 
     private String replaceAll(Pattern pattern, String replacement, String string) {
