+++ src/test/java/TestMultiLineMacros.java
@@ -1,30 +1,9 @@
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
-import org.junit.jupiter.params.provider.ValueSource;
-import org.tinylog.Logger;
-import org.variantsync.diffdetective.variation.DiffLinesLabel;
-import org.variantsync.diffdetective.variation.diff.VariationDiff;
-import org.variantsync.diffdetective.feature.CPPAnnotationParser;
-import org.variantsync.diffdetective.variation.diff.parse.VariationDiffParseOptions;
-import org.variantsync.diffdetective.variation.diff.parse.VariationDiffParser;
-import org.variantsync.diffdetective.variation.diff.serialize.LineGraphExportOptions;
-import org.variantsync.diffdetective.variation.diff.serialize.VariationDiffSerializeDebugData;
-import org.variantsync.diffdetective.variation.diff.serialize.GraphFormat;
-import org.variantsync.diffdetective.variation.diff.serialize.LineGraphExport;
-import org.variantsync.diffdetective.variation.diff.serialize.edgeformat.DefaultEdgeLabelFormat;
-import org.variantsync.diffdetective.variation.diff.serialize.nodeformat.DebugDiffNodeFormat;
-import org.variantsync.diffdetective.variation.diff.serialize.treeformat.CommitDiffVariationDiffLabelFormat;
 import org.variantsync.diffdetective.diff.result.DiffParseException;
-import org.variantsync.diffdetective.util.IO;
-import org.variantsync.diffdetective.util.StringUtils;
 
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-
-import java.io.BufferedReader;
 import java.io.IOException;
-import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.ArrayList;
 import java.util.stream.Stream;
 
 public class TestMultiLineMacros {
