+++ src/test/java/TreeDiffingTest.java
@@ -1,15 +1,14 @@
 import com.github.gumtreediff.matchers.Matcher;
 import com.github.gumtreediff.matchers.Matchers;
-
 import org.apache.commons.io.IOUtils;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.variantsync.diffdetective.diff.result.DiffParseException;
-import org.variantsync.diffdetective.feature.CPPAnnotationParser;
+import org.variantsync.diffdetective.feature.cpp.CPPAnnotationParser;
 import org.variantsync.diffdetective.util.IO;
 import org.variantsync.diffdetective.variation.DiffLinesLabel;
-import org.variantsync.diffdetective.variation.diff.construction.GumTreeDiff;
 import org.variantsync.diffdetective.variation.diff.VariationDiff;
+import org.variantsync.diffdetective.variation.diff.construction.GumTreeDiff;
 import org.variantsync.diffdetective.variation.diff.parse.VariationDiffParseOptions;
 import org.variantsync.diffdetective.variation.diff.parse.VariationDiffParser;
 import org.variantsync.diffdetective.variation.diff.serialize.Format;
@@ -21,19 +20,19 @@ import org.variantsync.diffdetective.variation.diff.serialize.nodeformat.FullNod
 import org.variantsync.diffdetective.variation.tree.VariationTree;
 import org.variantsync.diffdetective.variation.tree.source.LocalFileSource;
 
-import static org.junit.jupiter.api.Assertions.fail;
-
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.regex.Pattern;
 import java.util.stream.Stream;
 
+import static org.junit.jupiter.api.Assertions.fail;
 import static org.variantsync.diffdetective.variation.diff.Time.BEFORE;
 
 public class TreeDiffingTest {
     private final static Path testDir = Constants.RESOURCE_DIR.resolve("tree-diffing");
     private static Pattern expectedFileNameRegex = Pattern.compile("([^_]+)_([^_]+)_expected.lg");
+
     private static record TestCase(String basename, String matcherName, Matcher matcher) {
         public Path beforeEdit() {
             return testDir.resolve(String.format("%s.before", basename()));
@@ -58,20 +57,20 @@ public class TreeDiffingTest {
 
     private static Stream<TestCase> testCases() throws IOException {
         return Files
-            .list(testDir)
-            .mapMulti(((path, result) -> {
-                String filename = path.getFileName().toString();
-                var filenameMatcher = expectedFileNameRegex.matcher(filename);
-                if (filenameMatcher.matches()) {
-                    var treeMatcherName = filenameMatcher.group(2);
+                .list(testDir)
+                .mapMulti(((path, result) -> {
+                    String filename = path.getFileName().toString();
+                    var filenameMatcher = expectedFileNameRegex.matcher(filename);
+                    if (filenameMatcher.matches()) {
+                        var treeMatcherName = filenameMatcher.group(2);
 
-                    result.accept(new TestCase(
-                        filenameMatcher.group(1),
-                        treeMatcherName,
-                        Matchers.getInstance().getMatcher(treeMatcherName))
-                    );
-                }
-            }));
+                        result.accept(new TestCase(
+                                filenameMatcher.group(1),
+                                treeMatcherName,
+                                Matchers.getInstance().getMatcher(treeMatcherName))
+                        );
+                    }
+                }));
     }
 
     @ParameterizedTest
@@ -84,7 +83,7 @@ public class TreeDiffingTest {
 
         try (var output = IO.newBufferedOutputStream(testCase.actual())) {
             new LineGraphExporter<>(new Format<>(new FullNodeFormat(), new ChildOrderEdgeFormat<>()))
-                .exportVariationDiff(variationDiff, output);
+                    .exportVariationDiff(variationDiff, output);
         }
 
         try (
@@ -97,16 +96,16 @@ public class TreeDiffingTest {
             } else {
                 // Keep output files if the test failed
                 new TikzExporter<>(new Format<>(new FullNodeFormat(), new DefaultEdgeLabelFormat<>()))
-                    .exportFullLatexExample(variationDiff, testCase.visualisation());
+                        .exportFullLatexExample(variationDiff, testCase.visualisation());
                 fail(String.format(
-                    "The diff of %s and %s is not as expected. " +
-                    "Expected the content of %s but got the content of %s. " +
-                    "Note: A visualisation is available at %s",
-                    testCase.beforeEdit(),
-                    testCase.afterEdit(),
-                    testCase.expected(),
-                    testCase.actual(),
-                    testCase.visualisation()
+                        "The diff of %s and %s is not as expected. " +
+                                "Expected the content of %s but got the content of %s. " +
+                                "Note: A visualisation is available at %s",
+                        testCase.beforeEdit(),
+                        testCase.afterEdit(),
+                        testCase.expected(),
+                        testCase.actual(),
+                        testCase.visualisation()
                 ));
             }
         }
@@ -115,14 +114,14 @@ public class TreeDiffingTest {
     public VariationTree<DiffLinesLabel> parseVariationTree(Path filename) throws IOException, DiffParseException {
         try (var file = Files.newBufferedReader(filename)) {
             return new VariationTree<>(
-                VariationDiffParser.createVariationTree(
-                    file,
-                    new VariationDiffParseOptions(
-                        CPPAnnotationParser.Default,
-                        false,
-                        false)
-                ).getRoot().projection(BEFORE).toVariationTree(),
-                new LocalFileSource(filename)
+                    VariationDiffParser.createVariationTree(
+                            file,
+                            new VariationDiffParseOptions(
+                                    CPPAnnotationParser.Default,
+                                    false,
+                                    false)
+                    ).getRoot().projection(BEFORE).toVariationTree(),
+                    new LocalFileSource(filename)
             );
         }
     }
